1.improve and complete circle progress bar 2.converting all text objects to Upersian 3.add socket client class 4.change font styles

+++ Rehabilitation_Game/Assets/scripts/Aim.cs
@@ -59,7 +59,7 @@ public class Aim : MonoBehaviour
         movementRange = Screen.width * movement / 200 / screenWidthInUnit;
     }
 
-    public void setScale(int scale)
+    public void setScale(float scale)
     {
         transform.localScale = new Vector3(scale, scale);
     }
